requirements.txt
jupyter~=1.0.0
pip~=24.0
attrs~=23.1.0
wheel~=0.42.0
Pillow~=10.1.0
tornado~=6.3.3
Jinja2~=3.1.2
setuptools~=58.1.0
packaging~=23.2
pyparsing~=3.1.1
future~=0.18.3
numpy~=1.26.1
pywebview~=5.0.5
tensorflow~=2.16.1
Flask~=2.3.2